feat: Refactor image server to use new API structure; implement image upload and retrieval routes; enhance tunnel management with dedicated controllers
