+++ electron/imageServer/index.html
@@ -217,8 +217,8 @@
                     
                     return `
                         <div class="item" style="grid-column: span ${Math.min(gridSpan, 3)};">
-                            <img 
-                                src="/images/${image.id}" 
+                            <img
+                                src="/api/images/${image.id}" 
                                 alt="${image.name}" 
                                 loading="lazy"
                                 style="aspect-ratio: ${aspectRatio}"
